Rename layout.utils to layout.service
