+++ frontend/src/scripts/modules/layout/layout.service.ts
